Proto/web samples (#96)
Initial implementation of a patterns JSON format that we can later use to highlight an agent sample on a website. The contribution is inspired by serverlessland.com and associated GitHub process, based around a minimal format in a _template directory that can be consumed by downstream websites or other APIs.